Add detailed README and modularize generator
